requirements.txt
discord.py>=2.0.0
sqlalchemy>=1.4.0
python-dotenv>=0.19.0
